requirements.txt
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-filter==22.1

# jwtstuff
djangorestframework_simplejwt==5.2.0
django-axes==5.25.0



